Update project configuration and testing setup
- Add OpenAI and pytest dependencies to pyproject.toml
- Remove outdated run_test.py and replace with test_text_chat.py for improved testing structure
- Introduce development scripts for streamlined package management and testing
- Enhance DEVELOPMENT.md with comprehensive usage instructions for UV
