+++ pyproject.toml
@@ -6,3 +6,18 @@ requires-python = ">=3.11"
 dependencies = [
   "openai>=1.0.0"
 ]
+
+[project.optional-dependencies]
+test = [
+  "pytest>=7.0.0"
+]
+
+[tool.uv]
+package = true
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = "test_*.py"
+python_classes = "Test*"
+python_functions = "test_*"
+addopts = "-v --tb=short"
